Celestial Bodies

Potential Models

Jamie.NullPotential — Type

NullPotential

Empty potential model used as an indicator that no potential is defined.

See also: PointMassPotential

Jamie.PointMassPotential — Type

PointMassPotential(gm::T)

Simple potential model for a point mass (or centrobaric) body of gravitational parameter gm.

See also: NullPotential

Jamie.gravitational_parameter — Method

gravitational_parameter(::PointMassPotential)

Retrieve the gravitational parameter, GM, of the potential model. This is the standard 2-body μ value.

See also: PointMassPotential

Shape Models

Jamie.NullShapeModel — Type

NullShapeModel

Empty shape model used as an indicator that no shape is defined.

See also: SphericalShapeModel

Jamie.SphericalShapeModel — Type

SphericalShapeModel(radius::T)

Simple shape model describing a sphere of radius, radius.

See also: NullShapeModel

Jamie.mean_radius — Method

mean_radius(::SphericalShapeModel)

Retrieve the mean radius of the spherical shape model. Clearly for this model, the mean radius is simple equal to the radius, radius.

See also: NullShapeModel

The Celestial Body Structure

Jamie.CelestialBody — Type

CelestialBody{T, I <: Integer}(name::String, gm::T, spice_id::I, mean_radius::T)

Struct defining a celestial body (e.g. Moon, Planet, Asteroid).

See also: PointMassPotential, SphericalShapeModel

Jamie.name_string — Method

name_string(cb::CelestialBody)

Retrieve the name of the celestial body.

See also: CelestialBody, spice_identifier, potential_model, shape_model, mean_radius, gravitational_parameter

Jamie.spice_identifier — Function

spice_identifier(cb::CelestialBody)

Retrieve the integer SPICE identifier of the celestial body. More information on SPICE ID codes can be found here.

See also: CelestialBody, name_string, potential_model, shape_model, mean_radius, gravitational_parameter

Jamie.potential_model — Function

potential_model(::CelestialBody)

Retrieve the potential modlel of the body.

See also: CelestialBody, name_string, spice_identifier, shape_model, mean_radius, gravitational_parameter

Jamie.shape_model — Function

shape_model(::CelestialBody)

Retrieve the shape model of the body.

See also: CelestialBody, name_string, spice_identifier, potential_model, mean_radius, gravitational_parameter

Jamie.mean_radius — Method

mean_radius(::CelestialBody)

Retrieve the mean radius of the body if a non-null shape model is defined. The particular calculation of the mean radius is implemented by the shape model.

See also: SphericalShapeModel, CelestialBody, name_string, spice_identifier, potential_model, shape_model, gravitational_parameter

Jamie.gravitational_parameter — Method

gravitational_parameter(::CelestialBody)

Retrieve the gravitational parameter of the body if a non-null potential model is defined. The particular calculation of the gravitational parameter is implemented by the potential model.

See also: PointMassPotential, CelestialBody, name_string, spice_identifier, potential_model, shape_model, mean_radius,