Position-Velocity

Jamie.PositionVelocity — Type

PositionVelocity{T} <: FieldVector{6, T}

Struct representing the position-velocity vector. T designates the types of component in the vector.

See also: position, velocity, position_magnitude, position_direction, velocity_magnitude, velocity_direction

Base.position — Function

position(s)

Get the current position of a stream.

Examples

julia> io = IOBuffer("JuliaLang is a GitHub organization.");

julia> seek(io, 5);

julia> position(io)
5

julia> skip(io, 10);

julia> position(io)
15

julia> seekend(io);

julia> position(io)
35

position(::PositionVelocity)

Get the position components of the position-velocity vector.

See also: velocity, position_magnitude, position_direction, velocity_magnitude, velocity_direction

Jamie.position_magnitude — Function

position_magnitude(::PositionVelocity)

Return the magnitude of the position components of the position-velocity vector. This is equivalent to the distance from the origin of the position vector.

See also: position, velocity, position_direction, velocity_magnitude, velocity_direction

Jamie.position_direction — Function

position_direction(::PositionVelocity)

Return the unit vector parallel to the position sub-vector in the specified position-velocity vector.

See also: position, velocity, position_magnitude, velocity_magnitude, velocity_direction

Jamie.velocity — Function

velocity(::PositionVelocity)

Get the velocity components of the position-velocity vector.

See also: position, position_magnitude, position_direction, velocity_magnitude, velocity_direction

Jamie.velocity_magnitude — Function

velocity_magnitude(::PositionVelocity)

Return the magnitude of the velocity components of the position-velocity vector.

See also: position, velocity, position_magnitude, position_direction, velocity_direction

Jamie.velocity_direction — Function

velocity_direction(::PositionVelocity)

Return the unit vector parallel to the velocity sub-vector in the specified position-velocity vector.

See also: position, velocity, position_magnitude, position_direction, velocity_magnitude,