Utilities

Generating Points on a Sphere

Jamie.points_on_sphere — Function

points_on_sphere(n; [radius=1.0], [center=(0.0, 0.0, 0.0)])

Generate a set of points on a sphere approximately uniformly distributed via the Fibonacci lattice method. The points are returned as an Array of n SVectors of length 3. A specific radius and center can be optionally given as key-word arguments.